+++ examples/nsga2_zdt1.rs
@@ -17,7 +17,7 @@ use optirustic::core::builtin_problems::ZTD1Problem;
 ///
 /// Make sure to compile this in release mode to speed up the calculation:
 ///
-/// `cargo run --example nsga2_zdt1 --release`
+/// `cargo run --example nsga2_zdt1 -p optirustic --release`
 fn main() -> Result<(), Box<dyn Error>> {
     // Add log
     env_logger::builder().filter_level(LevelFilter::Info).init();